refactor: improve delete student

+++ src/main/java/com/services/ms/student/app/application/service/StudentService.java
@@ -46,7 +46,9 @@ public class StudentService implements StudentServicePort {
 
     @Override
     public void deleteById(Long id) {
-        persistencePort.deleteById(id);
+        if (persistencePort.findById(id).isEmpty()){
+            persistencePort.deleteById(id);
+        }
     }
     
 }
